Add section with prototype hardware.
The current hardware is listed in the Google doc, but unless we want to link that here if we ever open source the repo it's hard for people to know how to do DIY this. I also think that having the current version of the hardware alongside the code in a repo makes sense.